requirements.txt
asgiref==3.3.1
Django==3.1.6
gTTS==2.2.2
Pillow==8.1.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
gunicorn==20.0.4
whitenoise==5.0.1